Updated .graphqlrc.yaml to reflect correct schema location
I updated the `.graphqlrc.yaml` file to reflect the same schema location (URL) as is used in the `KitQLClient` file for consistency through the process and to avoid errors. I also added comments to educate new users on what the `schema` field and the URL in the client are used for. 

Also note that I was not able to get the walkthrough working with these changes alone - reference [issue #132](https://github.com/jycouet/kitql/issues/132) for further info